+++ Runtime/Tweens/EntityTween.cs
@@ -105,6 +105,55 @@ namespace Timespawn.EntityTween.Tweens
             parallelWriter.AddComponent(sortKey, entity, new TweenRotationCommand(tweenParams, start, end));
         }
 
+        public static void Scale(
+            EntityManager entityManager,
+            Entity entity,
+            float3 start, 
+            float3 end, 
+            float duration,
+            EaseDesc easeDesc = default,
+            bool isPingPong = false, 
+            int loopCount = 1)
+        {
+            AssertParams(easeDesc.Exponent, loopCount);
+
+            TweenParams tweenParams = new TweenParams(easeDesc.Type, easeDesc.Exponent, duration, isPingPong, loopCount);
+            entityManager.AddComponentData(entity, new TweenScaleCommand(tweenParams, start, end));
+        }
+
+        public static void Scale(
+            EntityCommandBuffer commandBuffer,
+            Entity entity,
+            float3 start, 
+            float3 end, 
+            float duration,
+            EaseDesc easeDesc = default,
+            bool isPingPong = false, 
+            int loopCount = 1)
+        {
+            AssertParams(easeDesc.Exponent, loopCount);
+
+            TweenParams tweenParams = new TweenParams(easeDesc.Type, easeDesc.Exponent, duration, isPingPong, loopCount);
+            commandBuffer.AddComponent(entity, new TweenScaleCommand(tweenParams, start, end));
+        }
+
+        public static void Scale(
+            EntityCommandBuffer.ParallelWriter parallelWriter,
+            int sortKey,
+            Entity entity,
+            float3 start, 
+            float3 end, 
+            float duration,
+            EaseDesc easeDesc = default,
+            bool isPingPong = false, 
+            int loopCount = 1)
+        {
+            AssertParams(easeDesc.Exponent, loopCount);
+
+            TweenParams tweenParams = new TweenParams(easeDesc.Type, easeDesc.Exponent, duration, isPingPong, loopCount);
+            parallelWriter.AddComponent(sortKey, entity, new TweenScaleCommand(tweenParams, start, end));
+        }
+
         private static void AssertParams(int easeExponent, int loopCount)
         {
             Debug.Assert(byte.MinValue <= easeExponent && easeExponent <= byte.MaxValue, $"Exponent of ease function should be between {byte.MinValue} - {byte.MaxValue}.");