+++ CHANGELOG.md
@@ -6,4 +6,4 @@
 
 - `Ease` utilities for interpolation
 - Tween API using `EntityManager`, `EntityCommandBuffer` and `ParallelWriter`
-- `Translation` and `Rotation` tween support
+- `Translation`, `Rotation` and `Scale` tween support